Rename magnetometerOptions.h to options_magnetometer.h in MPLAB-X. Reset options to their default settings.
