+++ Config/options_magnetometer.h
@@ -31,7 +31,7 @@
 // HMC5883L_SF is the SparkFun HMC5883L
 
 //#define HMC5843
-#define HMC5883L
+//#define HMC5883L
 //#define HMC5883L_SF
 
 // Define magneticDeclination to be the magnectic declination, in degrees, measured
@@ -62,7 +62,7 @@
 // For 3DRobotics mags, for MAG_DIRECT the mag mounts over the UDB, component side down.
 // For SparkFun HMC5883L, for MAG_DIRECT the mag mounts over the UDB, component side up.
 
-#define MAG_FORWARDS
+//#define MAG_FORWARDS
 //#define MAG_BACKWARDS
 //#define MAG_INVERTED
 //#define MAG_FLIPPED
