+++ MatrixPilot/MatrixPilot-udb5.X/nbproject/configurations.xml
@@ -21,7 +21,6 @@
         </logicalFolder>
         <itemPath>../../Config/airspeed_options.h</itemPath>
         <itemPath>../../Config/FSconfig.h</itemPath>
-        <itemPath>../../Config/magnetometerOptions.h</itemPath>
         <itemPath>../../Config/mavlink_options.h</itemPath>
         <itemPath>../../Config/nv_memory_options.h</itemPath>
         <itemPath>../../Config/options.h</itemPath>
@@ -31,6 +30,7 @@
         <itemPath>../../Config/servoMix_options.h</itemPath>
         <itemPath>../../Config/waypoints.h</itemPath>
         <itemPath>../../Config/flightplan-logo.h</itemPath>
+        <itemPath>../../Config/options_magnetometer.h</itemPath>
       </logicalFolder>
       <logicalFolder name="libDCM" displayName="libDCM" projectFiles="true">
         <itemPath>../../libDCM/dcmTypes.h</itemPath>
