+++ Tools/LedTest/Config/options_magnetometer.h
