+++ Tools/LedTest/LedTest-udb5.X/nbproject/configurations.xml
@@ -85,11 +85,11 @@
                      displayName="Tools/LedTest/Config"
                      projectFiles="true">
         <itemPath>../../../Tools/LedTest/Config/FSconfig.h</itemPath>
-        <itemPath>../../../Tools/LedTest/Config/magnetometerOptions.h</itemPath>
         <itemPath>../../../Tools/LedTest/Config/nv_memory_options.h</itemPath>
         <itemPath>../../../Tools/LedTest/Config/options.h</itemPath>
         <itemPath>../../../Tools/LedTest/Config/osd_config.h</itemPath>
         <itemPath>../../../Tools/LedTest/Config/ports_config.h</itemPath>
+        <itemPath>../../../Config/options_magnetometer.h</itemPath>
       </logicalFolder>
     </logicalFolder>
     <logicalFolder name="LinkerScript"
@@ -212,6 +212,7 @@
     <Elem>../../libVectorMatrix</Elem>
     <Elem>../../MatrixPilot</Elem>
     <Elem>../../MAVLink</Elem>
+    <Elem>../../../Config</Elem>
   </sourceRootList>
   <projectmakefile>Makefile</projectmakefile>
   <confs>
