+++ libDCM/mag_drift.c
@@ -23,7 +23,7 @@
 #include "mathlibNAV.h"
 #include "../libUDB/heartbeat.h"
 #include "../libUDB/magnetometer.h"
-#include "magnetometerOptions.h"
+#include "options_magnetometer.h"
 #include "mag_drift.h"
 #include "rmat.h"
